+++ Api_Retail.lua
@@ -266,6 +266,6 @@ end
 __Static__() __AutoCache__()
 function AshBlzSkinApi.UnitDebuffCanDispellColor()
     return AshBlzSkinApi.UnitDebuffCanDispell():Map(function(_, dType)
-        return DebuffTypeColor[dType or ""]
+        return GetDispellDebuffColor(dType or "")
     end)
 end